Added new macro system to reduce repeating in pattern matching, improved performance.
File type detection now happens through a procedural macro, reducing the
overall visual noise of the program and making it infinitely easy to add
new supported formats. To add new format, it is now not needed to write
anything but to add new member to central Format enum and give it
necessary parameters if needed. It is AT MOST three lines of code/file
format.

To store the amount of lines while processing the files, wclines now
employs a hash table from file format to overall amount of lines found
in files of this format. This remarkably simplifies accesing current
state of work while not harming preformance much.

Also, i started sketching ways to ignore binary files since it is
useless for you to know how much random sequences of bytes happened to
decode into a newline character in a .mp3 or .exe file.
Currently, only ELF (native Linux executable format) files are ignored
by doing a silly comparison between ELF signature and first bytes of a
file, but it is not the main purpose of this commit.